necessary folders and files
