+++ src/components/App/App.js
@@ -1,0 +1,34 @@
+import logo from "../../logo.svg";
+import "../App/App.css";
+import React, { useState, useEffect } from "react";
+import { Switch, Route } from "react-router-dom";
+import { useHistory } from "react-router-dom";
+import Header from "../Header/Header";
+import Main from "../Main/Main";
+import Footer from "../Footer/Footer";
+
+function App() {
+  return (
+    <div className="App">
+      {/* <header className="App-header">
+        <img src={logo} className="App-logo" alt="logo" />
+        <p>
+          Edit <code>src/App.js</code> and save to reload.
+        </p>
+        <a
+          className="App-link"
+          href="https://reactjs.org"
+          target="_blank"
+          rel="noopener noreferrer"
+        >
+          Learn React
+        </a>
+      </header> */}
+      <Header />
+      <Main />
+      <Footer />
+    </div>
+  );
+}
+
+export default App;
